+++ ConvexHullAlgorithm/ConvexHullAlgorithm/point.h
@@ -16,9 +16,9 @@ public:
 	virtual void print() = 0;
 
 	/* Get Values */
-	virtual float getX() = 0;
-	virtual float getY() = 0;
-	virtual float getZ() = 0;
+	virtual double getX() = 0;
+	virtual double getY() = 0;
+	virtual double getZ() = 0;
 
 	/* Get Type */
 	virtual PointType getType() = 0;
